Moved Header and Footer globals to a globals/ folder, consistent with collections
