+++ apps/pragmatic-papers/src/globals/Footer/Component.tsx
