+++ apps/pragmatic-papers/src/globals/Footer/RowLabel.tsx
