+++ apps/pragmatic-papers/src/globals/Footer/hooks/revalidateFooter.ts
