+++ apps/pragmatic-papers/src/globals/Header/hooks/revalidateHeader.ts
